chore(): rewmove dir 1
